First serious checkpoint
The game is nowhere near playable yet. Mostly, the code is the 'setup" code,
corresponding to the earliest lines of the BASIC code, with some elements
from subroutines later in the code brought in early as well. For example,
the routines that map numbers to star maps, or damage control, have been
modeled as idiomatic Ruby already.

See the README for the approach I'm taking, which leans toward better Ruby
rather than a more literal translation.